Split tests into files
