+++ test/gbmv.js
@@ -9,8 +9,9 @@ var assertCloseTo = require('./util/close-to');
 var constants = require('./util/constants');
 
 var gemv = require('../gemv');
+var gbmv = require('../gbmv');
 
-describe('GEMV (general matrix vector product)', function () {
+describe('GBMV (general banded matrix-vector product)', function () {
   var m = 10;
   var n = 15;
   var alpha = 0;
@@ -19,20 +20,19 @@ describe('GEMV (general matrix vector product)', function () {
   var matGen = new RandMatGen(seed, Float64Array);
   var A = ndarray(new Float64Array(m * n), [m, n]);
   var x = ndarray(new Float64Array(n), [n]);
+  var x0 = ndarray(new Float64Array(n), [n]);
+  var xn = ndarray(new Float64Array(n), [n]);
   var y = ndarray(new Float64Array(m), [m]);
   var y0 = ndarray(new Float64Array(m), [m]);
+  var B = ndarray(new Float64Array(n * n), [n, n]);
 
-  it('gemv', function () {
+  it('gbmv', function () {
     for (var t = 0; t < constants.NUM_TESTS; ++t) {
-      alpha = Math.random();
-      beta = Math.random();
       seed = matGen.setRandomSeed(36);
-      matGen.makeGeneralMatrix(m, n, A);
-      matGen.makeGeneralMatrix(1, n, x);
-
-      assert(gemv(alpha, A, x, beta, y));
-      assert(naiveGEMV(alpha, A, x, beta, y0));
-      assertCloseTo(y0, y, constants.TEST_TOLERANCE, 'Failure seed value: "' + seed + '".');
+      matGen.makeBandedMatrix(n, n, 3, 1, B);
+      gbmv(B, 3, 1, x0, x);
+      gemv(1, B, x0, 0, xn);
+      assertCloseTo(x, xn, constants.TEST_TOLERANCE, 'Failure seed value: "' + seed + '".');
     }
   });
 });
