+++ test/util/rand-matrix-gen.js
